New Instagram API
Instagram API changes – April 4, 2018

Due to some sudden changes that Instagram has made to their API on April 4th, 2018

User Feeds
It is still possible to display your own Instagram account, however, Instagram has removed the ability to display user feeds from other Instagram accounts which are not your own. You can only display the user feed of the account which is associated with your Access Token, but it is no longer possible to display someone else’s feed using your own Access Token.

Public Comments
It is possible to display comments on your own user feeds, but it is no longer possible to display comments on posts in hashtag or location feeds.

Avatars
It is no longer possible to display the avatars of users in hashtag or location feeds.